requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.2.20
django-celery-results==2.0.1
django-filter==2.4.0
djangorestframework==3.11.2
djangorestframework-camel-case==1.1.2
drf-yasg==1.17.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.3
kombu==4.6.11
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.3
packaging==19.0
pandas==0.24.2
psutil==5.7.2
psycopg2-binary==2.8.6
pygbif==0.2.0
pyparsing==2.4.1.1
python-dateutil==2.8.0
python-dwca-reader==0.14.0
pytz==2019.1
PyYAML==5.4
redis==3.2.1
requests==2.23.0
ruamel.yaml==0.15.99
Shapely==1.7.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
